Put in J2 term for Hamiltonian also.  Test results (storing the Hamiltonian as long double) J1=1.0 J2=1.600 Q=0.50 Lanczos: -21.3890818508  -20.3996627058
J1=0.50
J2=0.90
Q=1.20
Lanczos: -19.4199173175  -18.8778317143  

J1=0.0
J2=1.00
Q=1.00
Lanczos: -16.2181547489  -15.839022948

+++ trunk/GenHam.h
@@ -10,7 +10,7 @@ using namespace std;
 #include <blitz/array.h>
 BZ_USING_NAMESPACE(blitz)
 
-typedef float h_float;  //precision for Hamiltonian storage
+typedef long double h_float;  //precision for Hamiltonian storage
 
 class GENHAM{
